Extract src/panes.ts — the three spawners and a pane's life
Step 3 and the last of the bootstrap trim's three blocks, and the one PLAN
flagged as most interconnected. 362 lines out in eleven guarded ranges:
launch, launchShell, launchTask, requestLaunch, closeSession, setActive,
refreshSessionStats, refreshBranch, refreshBranches, scheduleDismiss,
renderHeader, syncStageButtons, activeProjectCtx, activeCwd,
openPlainTerminal, handToTerminal.

They belong together because spawn_claude / spawn_shell / spawn_task are three
backend entry points but one frontend shape — a Sess in the map, a .term-pane
in #terminals, a PTY behind it. renderHeader and syncStageButtons came too:
both read only which pane is active, and moving them killed two hooks the
alternative would have needed.

**One hook for the whole module** — setPanesRenderAll. Everything else it
needs was already below it after the earlier slices, which is the clearest
sign the ordering PLAN recorded was the right one.

A real constraint found and recorded rather than papered over: taskrun.ts's
three setters (setActive / closeSession / launchTask) could now become direct
imports and disappear — except inspector.ts imports taskrun, and panes imports
inspector for renderInspector, so that would close the loop
inspector -> taskrun -> panes -> inspector. Ground rule 3 forbids import
cycles, so the setters stay; three setters are the cheaper side of that trade.
Verified with a cycle check over the whole frontend graph: no cycles across 33
modules. PLAN says to re-run it after any slice that adds an import.

Verified mechanical: line-sorted against `git show HEAD:src/main.ts` the
relocated block is identical modulo the export keywords. (First run of that
check compared a window three lines short and reported a difference that
wasn't one — re-derived the boundary by grep before believing it.)

main.ts 1,092 -> 730 lines. Gates: tsc --noEmit clean (exit 0), 368 vitest
pass. No Rust touched.
